# share/htdocs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  api.ActivityItem:
    properties:
      description:
        type: string
      timestamp:
        type: string
      type:
        type: string
    type: object
...
  api.AgentStats:
    properties:
      active:
        type: integer
      total:
        type: integer
    type: object
...
  api.DashboardStatsResponse:
    properties:
      agent_stats:
        $ref: '#/definitions/api.AgentStats'
      issue_stats:
        $ref: '#/definitions/api.IssueStats'
      recent_activity:
        items:
          $ref: '#/definitions/api.ActivityItem'
        type: array
      user_count:
        type: integer
      workspace_count:
        type: integer
    type: object
...
  api.ErrorResponse:
    description: Error information
    properties:
      error:
        example: Invalid request
        type: string
    type: object
...
  api.IssueStats:
    properties:
      by_priority:
        additionalProperties:
          type: integer
        type: object
      by_status:
        additionalProperties:
          type: integer
        type: object
      total:
        type: integer
    type: object
  api.LoginRequest:
    description: Login credentials
    properties:
      password:
        example: password123
        type: string
      username:
        example: admin
        type: string
    type: object
  api.LoginResponse:
    description: Authentication response with session token
    properties:
      expires_at:
        example: "2025-01-01T12:00:00Z"
        type: string
      token:
        example: token_abc123...
        type: string
      user:
        $ref: '#/definitions/api.UserInfo'
    type: object
...
  api.RefreshResponse:
    description: Refreshed session information
    properties:
      expires_at:
        example: "2025-01-01T14:00:00Z"
        type: string
      token:
        example: token_xyz789...
        type: string
    type: object
...
  api.StatusResponse:
    description: Status message
    properties:
      status:
        example: ok
        type: string
    type: object
...
  api.UserInfo:
    description: User details
    properties:
      id:
        example: entity_user_admin
        type: string
      roles:
        example:
        - admin
        - user
        items:
          type: string
        type: array
      username:
        example: admin
        type: string
    type: object
...
host: localhost:8085
info:
  contact:
    email: support@entitydb.io
    name: EntityDB Support
  description: A temporal database with pure entity-based architecture
  license:
    name: MIT
    url: https://github.com/osakka/entitydb/blob/main/LICENSE
  termsOfService: https://github.com/osakka/entitydb
  title: EntityDB API
  version: 2.12.0
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with username and password using relationship-based
        security
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Authenticate user
      tags:
      - authentication
  /api/v1/auth/logout:
    post:
      consumes:
      - application/json
      description: Logout user by invalidating the current session
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - authentication
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: Refresh the current session token to extend expiration
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.LoginResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Refresh session token
      tags:
      - authentication
  /api/v1/auth/whoami:
    get:
      consumes:
      - application/json
      description: Get information about the currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.UserInfo'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user information
      tags:
      - authentication
...
  /api/v1/dashboard/stats:
    get:
      consumes:
      - application/json
      description: Get dashboard statistics and metrics
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.DashboardStatsResponse'
      security:
      - BearerAuth: []
      summary: Dashboard statistics
      tags:
      - dashboard
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and receive session token
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.LoginResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Login
      tags:
      - auth
  /auth/logout:
    post:
      description: Invalidate the current session
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.StatusResponse'
      security:
      - BearerAuth: []
      summary: Logout
      tags:
      - auth
  /auth/refresh:
    post:
      description: Refresh the session token expiration
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.RefreshResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Refresh Token
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: 'Bearer token authentication. Example: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"